Update links.md
Made a tiny change to trigger a push after changing the settings for GitHub Pages.

Per: https://docusaurus.io/docs/deployment#triggering-deployment-with-github-actions

(Site not deployed properly?)

Under your repo's Settings > Pages, make sure the "Source" (which is the source for the deployment files, not "source" as in our terminology) is set to "gh-pages" + "/ (root)".

That source setting was previously set to "Github Actions" rather than "Deploy from a Branch"